+++ realm/realm-annotations-processor/src/test/resources/io/realm/SimpleRealmProxy.java
@@ -29,10 +29,10 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 public class SimpleRealmProxy extends some.test.Simple
-    implements RealmObjectProxy, SimpleRealmProxyInterface {
+        implements RealmObjectProxy, SimpleRealmProxyInterface {
 
     static final class SimpleColumnInfo extends ColumnInfo
-        implements Cloneable {
+            implements Cloneable {
 
         public long nameIndex;
         public long ageIndex;
@@ -76,6 +76,7 @@ public class SimpleRealmProxy extends some.test.Simple
         proxyState.setConstructionFinished();
     }
 
+    @Override
     public void realm$injectObjectContext() {
         if (this.proxyState != null) {
             return;
@@ -89,12 +90,14 @@ public class SimpleRealmProxy extends some.test.Simple
         proxyState.setExcludeFields$realm(context.getExcludeFields());
     }
 
+    @Override
     @SuppressWarnings("cast")
     public String realmGet$name() {
         proxyState.getRealm$realm().checkIfValid();
         return (java.lang.String) proxyState.getRow$realm().getString(columnInfo.nameIndex);
     }
 
+    @Override
     public void realmSet$name(String value) {
         if (proxyState.isUnderConstruction()) {
             if (!proxyState.getAcceptDefaultValue$realm()) {
@@ -117,12 +120,14 @@ public class SimpleRealmProxy extends some.test.Simple
         proxyState.getRow$realm().setString(columnInfo.nameIndex, value);
     }
 
+    @Override
     @SuppressWarnings("cast")
     public int realmGet$age() {
         proxyState.getRealm$realm().checkIfValid();
         return (int) proxyState.getRow$realm().getLong(columnInfo.ageIndex);
     }
 
+    @Override
     public void realmSet$age(int value) {
         if (proxyState.isUnderConstruction()) {
             if (!proxyState.getAcceptDefaultValue$realm()) {
@@ -217,7 +222,7 @@ public class SimpleRealmProxy extends some.test.Simple
 
     @SuppressWarnings("cast")
     public static some.test.Simple createOrUpdateUsingJsonObject(Realm realm, JSONObject json, boolean update)
-        throws JSONException {
+            throws JSONException {
         final List<String> excludeFields = Collections.<String> emptyList();
         some.test.Simple obj = realm.createObjectInternal(some.test.Simple.class, true, excludeFields);
         if (json.has("name")) {
@@ -240,7 +245,7 @@ public class SimpleRealmProxy extends some.test.Simple
     @SuppressWarnings("cast")
     @TargetApi(Build.VERSION_CODES.HONEYCOMB)
     public static some.test.Simple createUsingJsonStream(Realm realm, JsonReader reader)
-        throws IOException {
+            throws IOException {
         some.test.Simple obj = new some.test.Simple();
         reader.beginObject();
         while (reader.hasNext()) {
@@ -407,7 +412,7 @@ public class SimpleRealmProxy extends some.test.Simple
     }
 
     @Override
-    public ProxyState realmGet$proxyState() {
+    public ProxyState<?> realmGet$proxyState() {
         return proxyState;
     }