+++ realm/realm-library/src/main/cpp/io_realm_SyncManager.cpp
@@ -14,32 +14,20 @@
  * limitations under the License.
  */
 
-#include <jni.h>
-
-#include <chrono>
-#include <functional>
-#include <mutex>
-#include <vector>
-
-#include <realm/group_shared.hpp>
-#include <realm/sync/history.hpp>
-#include <realm/sync/client.hpp>
-
 #include "io_realm_SyncManager.h"
 
-#include "object-store/src/sync/sync_manager.hpp"
+#include <realm/group_shared.hpp>
 
-#include "binding_callback_thread_observer.hpp"
+#include <sync/sync_manager.hpp>
+#include <sync/sync_session.hpp>
+#include <binding_callback_thread_observer.hpp>
+
 #include "util.hpp"
-
 #include "jni_util/jni_utils.hpp"
 #include "jni_util/java_method.hpp"
 
 using namespace realm;
-using namespace realm::sync;
 using namespace realm::jni_util;
-
-std::unique_ptr<Client> sync_client;
 
 struct AndroidClientListener : public realm::BindingCallbackThreadObserver {
 
@@ -58,28 +46,18 @@ struct AndroidClientListener : public realm::BindingCallbackThreadObserver {
     }
 } s_client_thread_listener;
 
-JNIEXPORT void JNICALL Java_io_realm_SyncManager_nativeInitializeSyncClient(JNIEnv* env, jclass)
-{
-    TR_ENTER()
-    if (sync_client) {
-        return;
+struct AndroidSyncLoggerFactory : public realm::SyncLoggerFactory {
+    std::unique_ptr<util::Logger> make_logger(Logger::Level level) override
+    {
+        auto logger = std::make_unique<CoreLoggerBridge>(std::string("REALM_SYNC"));
+        logger->set_level_threshold(level);
+        // Cast to std::unique_ptr<util::Logger>
+        return std::move(logger);
     }
-
-    try {
-        // Setup SyncManager
-        g_binding_callback_thread_observer = &s_client_thread_listener;
-
-        // Create SyncClient
-        sync::Client::Config config;
-        config.logger = &CoreLoggerBridge::shared();
-        sync_client = std::make_unique<Client>(std::move(config)); // Throws
-    }
-    CATCH_STD()
-}
+} s_sync_logger_factory;
 
 JNIEXPORT void JNICALL Java_io_realm_SyncManager_nativeReset(JNIEnv* env, jclass)
 {
-
     TR_ENTER()
     try {
         SyncManager::shared().reset_for_testing();
@@ -87,12 +65,38 @@ JNIEXPORT void JNICALL Java_io_realm_SyncManager_nativeReset(JNIEnv* env, jclass
     CATCH_STD()
 }
 
-JNIEXPORT void JNICALL Java_io_realm_SyncManager_nativeConfigureMetaDataSystem(JNIEnv* env, jclass, jstring baseFile)
+JNIEXPORT void JNICALL Java_io_realm_SyncManager_nativeInitializeSyncManager(JNIEnv* env, jclass, jstring sync_base_dir)
 {
     TR_ENTER()
     try {
-        JStringAccessor base_file_path(env, baseFile); // throws
+        JStringAccessor base_file_path(env, sync_base_dir); // throws
         SyncManager::shared().configure_file_system(base_file_path, SyncManager::MetadataMode::NoEncryption);
+
+        // Register Sync Client thread start/stop callback
+        g_binding_callback_thread_observer = &s_client_thread_listener;
+
+        // init logger
+        SyncManager::shared().set_logger_factory(s_sync_logger_factory);
+    }
+    CATCH_STD()
+}
+
+JNIEXPORT void JNICALL Java_io_realm_SyncManager_nativeSimulateSyncError(JNIEnv* env, jclass, jstring local_realm_path,
+                                                                         jint err_code, jstring err_message,
+                                                                         jboolean is_fatal)
+{
+    TR_ENTER()
+    try {
+        JStringAccessor path(env, local_realm_path);
+        JStringAccessor message(env, err_message);
+
+        auto session = SyncManager::shared().get_existing_active_session(path);
+        if (!session) {
+            ThrowException(env, IllegalArgument, concat_stringdata("Session not found: ", path));
+            return;
+        }
+        std::error_code code = std::error_code{static_cast<int>(err_code), realm::sync::protocol_error_category()};
+        SyncSession::OnlyForTesting::handle_error(*session, {code, std::string(message), to_bool(is_fatal)});
     }
     CATCH_STD()
 }