+++ realm/realm-library/src/main/java/io/realm/internal/ObjectServerFacade.java
@@ -18,6 +18,8 @@ package io.realm.internal;
 
 import android.content.Context;
 
+import java.lang.reflect.InvocationTargetException;
+
 import io.realm.RealmConfiguration;
 import io.realm.exceptions.RealmException;
 
@@ -34,12 +36,17 @@ public class ObjectServerFacade {
         //noinspection TryWithIdenticalCatches
         try {
             Class syncFacadeClass = Class.forName("io.realm.internal.SyncObjectServerFacade");
-            syncFacade = (ObjectServerFacade) syncFacadeClass.newInstance();
+            //noinspection unchecked
+            syncFacade = (ObjectServerFacade) syncFacadeClass.getDeclaredConstructor().newInstance();
         } catch (ClassNotFoundException ignored) {
         } catch (InstantiationException e) {
             throw new RealmException("Failed to init SyncObjectServerFacade", e);
         } catch (IllegalAccessException e) {
             throw new RealmException("Failed to init SyncObjectServerFacade", e);
+        } catch (NoSuchMethodException e) {
+            throw new RealmException("Failed to init SyncObjectServerFacade", e);
+        } catch (InvocationTargetException e) {
+            throw new RealmException("Failed to init SyncObjectServerFacade", e.getTargetException());
         }
     }