+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/AuthTests.java
@@ -6,6 +6,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import io.realm.ClientResetHandler;
 import io.realm.ErrorCode;
 import io.realm.ObjectServerError;
 import io.realm.Realm;
@@ -72,6 +73,11 @@ public class AuthTests extends BaseIntegrationTest {
                             @Override
                             public void onError(SyncSession session, ObjectServerError error) {
                                 fail("Session failed: " + error);
+                            }
+
+                            @Override
+                            public void onClientResetRequired(SyncSession session, ClientResetHandler handler) {
+                                fail("Client Reset");
                             }
                         })
                         .build();